outsource Requires code to package-specific files
