+++ src/ExponentialMap.jl
@@ -77,8 +77,6 @@ end
 function load_expokit_sparsematrixexp()
 return quote
 
-using .Expokit: expmv
-
 function get_column(spmexp::SparseMatrixExp{N}, j::Int)::Vector{N} where {N}
     n = size(spmexp, 1)
     aux = zeros(N, n)
@@ -230,8 +228,6 @@ end
 
 function load_expokit_exponentialmap()
 return quote
-
-using .Expokit: expmv
 
 """
     σ(d::AbstractVector{N}, em::ExponentialMap{N}) where {N<:Real}
@@ -478,8 +474,6 @@ end
 
 function load_expokit_exponentialprojectionmap()
 return quote
-
-using .Expokit: expmv
 
 """
     σ(d::AbstractVector{N},
