+++ src/init_Expokit.jl
@@ -1,0 +1,7 @@
+eval(quote
+    using .Expokit: expmv
+end)
+
+eval(load_expokit_sparsematrixexp())
+eval(load_expokit_exponentialmap())
+eval(load_expokit_exponentialprojectionmap())
